+++ app/src/Navigation/MainNavigation/MainNavigationStyles/MainStackHeaderStyle.js
